add user dao implement and DateTimeUtil

+++ src/main/java/com/qfg/ctu/servlet/rest/pojos/RestStatics.java
@@ -4,6 +4,6 @@ package com.qfg.ctu.servlet.rest.pojos;
  * Created by rbtq on 7/31/16.
  */
 public class RestStatics {
-    public int curMonth;
-    public int curDay;
+    public long curMonth;
+    public long curDay;
 }
